migrate to arrow functions

+++ packages/sdk/src/mcp/agents/api.ts
@@ -7,7 +7,7 @@ import {
 } from "../../index.ts";
 import {
   assertHasWorkspace,
-  canAccessWorkspaceResource,
+  assertWorkspaceResourceAccess,
 } from "../assertions.ts";
 import { AppContext, createTool } from "../context.ts";
 import { InternalServerError, NotFoundError } from "../index.ts";
@@ -58,9 +58,11 @@ export const listAgents = createTool({
   name: "AGENTS_LIST",
   description: "List all agents",
   inputSchema: z.object({}),
-  canAccess: canAccessWorkspaceResource,
-  handler: async (_, c) => {
+  handler: async (_, c, { name }) => {
     assertHasWorkspace(c);
+
+    await assertWorkspaceResourceAccess(name, {}, c)
+      .then(() => c.resourceAccess.grant());
 
     const { data, error } = await c.db
       .from("deco_chat_agents")
@@ -92,44 +94,34 @@ export const getAgent = createTool({
   name: "AGENTS_GET",
   description: "Get an agent by id",
   inputSchema: z.object({ id: z.string() }),
-  async canAccess(name, props, c) {
-    const hasAccess = await canAccessWorkspaceResource(name, props, c)
-      .catch(() => false);
-
-    if (hasAccess) {
-      return true;
-    }
-
-    const { data: agentData } = await c.db
-      .from("deco_chat_agents")
-      .select("visibility")
-      .eq("workspace", c.workspace!.value)
-      .eq("id", props.id)
-      .single();
-
-    // TODO: implement this using authorization system
-    if (agentData?.visibility === "PUBLIC") {
-      return true;
-    }
-
-    return false;
-  },
-  handler: async ({ id }, c) => {
+  handler: async ({ id }, c, { name }) => {
     assertHasWorkspace(c);
 
-    const { data, error } = id in WELL_KNOWN_AGENTS
-      ? {
-        data: WELL_KNOWN_AGENTS[id as keyof typeof WELL_KNOWN_AGENTS],
-        error: null,
-      }
-      : await c.db
-        .from("deco_chat_agents")
-        .select("*")
-        .eq("id", id)
-        .single();
+    c.resourceAccess.grant();
+
+    const [canAccess, { data, error }] = await Promise.all([
+      assertWorkspaceResourceAccess(name, { id }, c)
+        .then(() => true)
+        .catch(() => false),
+      id in WELL_KNOWN_AGENTS
+        ? Promise.resolve({
+          data: WELL_KNOWN_AGENTS[id as keyof typeof WELL_KNOWN_AGENTS],
+          error: null,
+        })
+        : c.db
+          .from("deco_chat_agents")
+          .select("*")
+          .eq("workspace", c.workspace!.value)
+          .eq("id", id)
+          .single(),
+    ]);
 
     if ((error && error.code == NO_DATA_ERROR) || !data) {
       throw new NotFoundError(id);
+    }
+
+    if (data.visibility !== "PUBLIC" && !canAccess) {
+      throw new NotFoundError(`Agent ${id} not found`);
     }
 
     if (error) {
@@ -144,9 +136,11 @@ export const createAgent = createTool({
   name: "AGENTS_CREATE",
   description: "Create a new agent",
   inputSchema: AgentSchema.partial(),
-  canAccess: canAccessWorkspaceResource,
-  handler: async (agent, c) => {
+  handler: async (agent, c, { name }) => {
     assertHasWorkspace(c);
+
+    await assertWorkspaceResourceAccess(name, agent, c)
+      .then(() => c.resourceAccess.grant());
 
     const [{ data, error }] = await Promise.all([
       c.db
@@ -176,9 +170,12 @@ export const updateAgent = createTool({
     id: z.string(),
     agent: AgentSchema.partial(),
   }),
-  canAccess: canAccessWorkspaceResource,
-  handler: async ({ id, agent }, c) => {
+  handler: async ({ id, agent }, c, { name }) => {
     assertHasWorkspace(c);
+
+    await assertWorkspaceResourceAccess(name, { id, agent }, c)
+      .then(() => c.resourceAccess.grant());
+
     const { data, error } = await c.db
       .from("deco_chat_agents")
       .update({ ...agent, id, workspace: c.workspace.value })
@@ -202,9 +199,12 @@ export const deleteAgent = createTool({
   name: "AGENTS_DELETE",
   description: "Delete an agent by id",
   inputSchema: z.object({ id: z.string() }),
-  canAccess: canAccessWorkspaceResource,
-  handler: async ({ id }, c) => {
+  handler: async ({ id }, c, { name }) => {
     assertHasWorkspace(c);
+
+    await assertWorkspaceResourceAccess(name, { id }, c)
+      .then(() => c.resourceAccess.grant());
+
     const { error } = await c.db
       .from("deco_chat_agents")
       .delete()
